Fixup `scalac` for #13061.
[ci skip-rust]
[ci skip-build-wheels]
